requirements.txt
Babel==2.2.0
behave==1.2.5
behave-django==0.3.0
binaryornot==0.4.0
CASWSGIMiddleware==1.1.0
chardet==2.3.0
click==6.2
cookiecutter==1.3.0
Django==1.8.1
django-adaptors==0.2.5
django-admin-bootstrapped==2.5.6
django-cas-client==1.2.0
django-cas-ng==3.4.2
django-crispy-forms==1.5.2
django-environ==0.4.0
django-filter==0.11.0
django-widget-tweaks==1.4.1
djangorestframework==3.2.4
factory-boy==2.6.1
fake-factory==0.5.7
future==0.15.2
httpie==0.9.2
Jinja2==2.7.3
MarkupSafe==0.23
mysqlclient==1.3.6
nose==1.3.7
parse==1.6.6
parse-type==0.3.4
pyasn1==0.1.9
Pygments==2.0.2
PyMySQL==0.6.7
python-dateutil==2.5.2
pytz==2016.3
requests==2.8.1
rsa==3.2
ruamel.base==1.0.0
ruamel.yaml==0.10.12
selenium==2.53.1
six==1.10.0
whichcraft==0.1.1
xlwt==1.0.0
